+++ kbase-extension/static/kbase/js/widgets/narrative_core/kbaseNarrativeAppPanel.js
@@ -17,6 +17,7 @@ define([
     'narrativeConfig',
     'util/display',
     'util/bootstrapDialog',
+    'util/bootstrapSearch',
     'util/icon',
     'text!kbase/templates/beta_warning_body.html',
     'yaml!ext_components/kbase-ui-plugin-catalog/src/plugin/modules/data/categories.yml',
@@ -28,7 +29,6 @@ define([
     'uuid',
     'narrative_core/catalog/kbaseCatalogBrowser',
     'kbase/js/widgets/narrative_core/kbaseAppCard',
-    'util/bootstrapAlert',
     'common/runtime',
     'kbaseNarrative',
     'catalog-client-api',
@@ -42,6 +42,7 @@ define([
     Config,
     DisplayUtil,
     BootstrapDialog,
+    BootstrapSearch,
     Icon,
     BetaWarningTemplate,
     Categories,
@@ -52,7 +53,6 @@ define([
     Uuid,
     KBaseCatalogBrowser,
     kbaseAppCard,
-    BootstrapAlert,
     Runtime
 ) {
     'use strict';
@@ -92,22 +92,7 @@ define([
 
             this.icon_colors = Config.get('icons').colors;
 
-            this.$searchDiv = $('<div>')
-                .addClass('input-group')
-                .css({ 'margin-bottom': '3px' })
-                .hide();
-
-            this.$searchInput = $('<input type="text">')
-                .addClass('form-control')
-                .attr('Placeholder', 'Search apps')
-                .on('input', function() {
-                    self.refreshPanel(this);
-                })
-                .on('keyup', function(e) {
-                    if (e.keyCode === 27) {
-                        self.$searchDiv.toggle({ effect: 'blind', duration: 'fast' });
-                    }
-                });
+            this.$searchDiv = $('<div>').hide();
 
             this.$numHiddenSpan = $('<span>0</span>');
             this.$showHideSpan = $('<span>show</span>');
@@ -124,20 +109,6 @@ define([
                     this.$showHideSpan.text(curText === 'show' ? 'hide' : 'show');
                 }, this));
 
-            var $clearSearchBtn = $('<span>')
-                .addClass('input-group-addon btn btn-default kb-method-search-clear')
-                .attr('type', 'button')
-                .append($('<span class="fa fa-times">'))
-                .click(
-                    $.proxy(function() {
-                        this.$searchInput.val('');
-                        this.$searchInput.trigger('input');
-                    }, this)
-                );
-
-            this.$searchDiv.append(this.$searchInput)
-                .append($clearSearchBtn);
-
             // placeholder for apps and methods once they're loaded.
             this.$methodList = $('<div>')
                 .css({
@@ -153,6 +124,13 @@ define([
                     .append(this.$searchDiv)
                     .append(this.$toggleHiddenDiv))
                 .append(this.$methodList);
+
+            this.bsSearch = new BootstrapSearch(this.$searchDiv, {
+                inputFunction: function(e) {
+                    self.refreshPanel(this);
+                },
+                placeholder: 'Search apps'
+            });
 
             // The 'loading' panel should just have a spinning gif in it.
             this.$loadingPanel = $('<div>')
@@ -175,23 +153,17 @@ define([
                 .append(this.$loadingPanel)
                 .append(this.$errorPanel));
 
-            $(document).on('filterMethods.Narrative',
-                $.proxy(function(e, filterString) {
-                    if (filterString) {
-                        this.$searchDiv.show({ effect: 'blind', duration: 'fast' });
-                        this.$searchInput.val(filterString);
-                        this.$searchInput.trigger('input');
-                    }
-                }, this)
-            );
+            $(document).on('filterMethods.Narrative', function(e, filterString) {
+                if (filterString) {
+                    this.$searchDiv.show({ effect: 'blind', duration: 'fast' });
+                    this.bsSearch.val(filterString);
+                }
+            }.bind(this));
 
-            $(document).on('removeFilterMethods.Narrative',
-                $.proxy(function() {
-                    this.$searchDiv.toggle({ effect: 'blind', duration: 'fast' });
-                    this.$searchInput.val('');
-                    this.$searchInput.trigger('input');
-                }, this)
-            );
+            $(document).on('removeFilterMethods.Narrative', function() {
+                this.$searchDiv.toggle({ effect: 'blind', duration: 'fast' });
+                this.bsSearch.val('');
+            }.bind(this));
 
             /* 'request' should be expected to be an object like this:
              * {
@@ -279,14 +251,14 @@ define([
                 .click(function() {
                     this.$searchDiv.toggle({ effect: 'blind', duration: 'fast' });
                     var new_height = this.$methodList.height();
-                    if(this.app_offset){
+                    if (this.app_offset) {
                         new_height = (new_height - 40) + 'px';
-                    }else{
+                    } else {
                         new_height = (new_height + 40) + 'px';
                     }
                     this.$methodList.css('height', new_height);
                     this.app_offset = !this.app_offset;
-                    this.$searchInput.focus();
+                    this.bsSearch.focus();
                 }.bind(this)));
 
             // Refresh button
@@ -585,10 +557,11 @@ define([
          */
         triggerApp: function(app, tag, parameters) {
             if (Jupyter.narrative.narrController.uiModeIs('view')) {
-                new BootstrapAlert({
+                new BootstrapDialog({
                     type: 'warning',
                     title: 'Warning',
-                    body: 'Read-only Narrative -- you may not add apps to this Narrative'
+                    body: 'Read-only Narrative -- you may not add apps to this Narrative',
+                    alertOnly: true
                 });
                 // alert('Read-only Narrative -- may not add apps to this Narrative');
                 return;
@@ -750,7 +723,7 @@ define([
 
         refreshPanel: function() {
             var panelStyle = this.currentPanelStyle,
-                filterString = this.$searchInput.val(),
+                filterString = this.bsSearch.val(),
                 appSet = this.renderedApps,
                 self = this,
                 $appPanel = $('<div>');
@@ -826,12 +799,12 @@ define([
             };
 
             var withCategories = function(a,b) {
-              var catA = (Categories.categories[a] || a).toLowerCase();
-              var catB = (Categories.categories[b] || b).toLowerCase();
-                   if (catA < catB) { return -1}
-              else if (catA > catB) { return  1}
-              else                  { return  0}
-            }
+                var catA = (Categories.categories[a] || a).toLowerCase();
+                var catB = (Categories.categories[b] || b).toLowerCase();
+                if (catA < catB) { return -1; }
+                else if (catA > catB) { return  1; }
+                else                  { return  0; }
+            };
 
             var buildAccordionPanel = function(style) {
                 /* first, get elements in order like this: