+++ src/biokbase/narrative/tests/test_widgetmanager.py
@@ -5,6 +5,7 @@ import IPython
 import mock
 import os
 from util import TestConfig
+from narrative_mock.mockclients import get_mock_client
 
 """
 Tests for the WidgetManager class
@@ -14,19 +15,13 @@ __author__ = 'Bill Riehl <wjriehl@lbl.gov>'
 
 class WidgetManagerTestCase(unittest.TestCase):
     @classmethod
-    @mock.patch('biokbase.narrative.widgetmanager.SpecManager')
-    def setUpClass(self, mock_sm):
+    def setUpClass(self):
         config = TestConfig()
         os.environ['KB_WORKSPACE_ID'] = '12345'  # That's the same workspace as my luggage!
-        specs_list = config.load_json_file(config.get('specs', 'app_specs_file'))
-        specs_dict = dict()
-        for s in specs_list:
-            specs_dict[s['info']['id']] = s
-        mock_sm.return_value.app_specs = {
-            'release': specs_dict,
-            'beta': specs_dict,
-            'dev': specs_dict
-        }
+        app_specs_list = config.load_json_file(config.get('specs', 'app_specs_file'))
+        app_specs_dict = dict()
+        for s in app_specs_list:
+            app_specs_dict[s['info']['id']] = s
         self.wm = WidgetManager()
         self.good_widget = "kbaseTabTable"
         self.bad_widget = "notAWidget"
@@ -58,14 +53,20 @@ class WidgetManagerTestCase(unittest.TestCase):
         self.assertIsInstance(
             self.wm.show_output_widget(
                 self.good_widget,
-                {'obj': 'TestObject'}
+                {'obj': 'TestObject'},
+                upas={'obj': '1/2/3'}
             ),
             IPython.core.display.Javascript
         )
 
     def test_show_output_widget_bad(self):
         with self.assertRaises(ValueError):
-            self.wm.show_output_widget(self.bad_widget, {'bad': 'inputs'}, check_widget=True)
+            self.wm.show_output_widget(
+                self.bad_widget,
+                {'bad': 'inputs'},
+                upas={'bad': '1/2/3'},
+                check_widget=True
+            )
 
     def test_show_external_widget(self):
         widget = self.wm.show_external_widget(
@@ -80,6 +81,58 @@ class WidgetManagerTestCase(unittest.TestCase):
                                               {},
                                               auth_required=True)
         self.assertIsInstance(widget, IPython.core.display.Javascript)
+
+    @mock.patch('biokbase.narrative.widgetmanager.clients.get', get_mock_client)
+    def test_show_data_cell(self):
+        """
+        Tests - should do the following:
+            def show_data_widget(self, upa, title=None, cell_id=None, tag="release"):
+        fail message with no upa
+        fail message with malformed upa
+        shouldn't care what title or cell_id are, but should test to make sure they wind up in
+            output code properly
+        fail if type spec'd app isn't present for some tag
+        otherwise, succeed and produce JS code.
+
+        test mocks.
+        """
+        js_obj = self.wm.show_data_widget("18836/5/1", "some title", "no_id")
+        print(js_obj.data)
+        self.assertIsValidCellCode(js_obj, {}, "viewer", "kbaseGenomeView", "no_id", "some title")
+
+    @mock.patch('biokbase.narrative.widgetmanager.clients.get', get_mock_client)
+    def test_infer_upas(self):
+        test_result_upa = "18836/5/1"
+        upas = self.wm.infer_upas("testCrazyExample", {
+            "obj_id1": 1,
+            "obj_id2": 2,
+            "obj_name1": "foo",
+            "obj_name2": "bar",
+            "obj_names": ["a", "b", "c"],
+            "obj_ref1": "1/2/3",
+            "obj_ref2": "4/5/6",
+            "obj_refs": ["7/8/9", "0/1/2"],
+            "ws_name": "some_ws",
+            "extra_param": "extra_value",
+            "other_extra_param": 0
+        })
+        self.assertEquals(upas['obj_id1'], test_result_upa)
+        self.assertEquals(upas['obj_id2'], test_result_upa)
+        self.assertEquals(upas['obj_name1'], test_result_upa)
+        self.assertEquals(upas['obj_name2'], test_result_upa)
+        self.assertEquals(upas['obj_names'], [test_result_upa]*3)
+        self.assertEquals(upas['obj_ref1'], "1/2/3")
+        self.assertEquals(upas['obj_ref2'], "4/5/6")
+        self.assertEquals(upas['obj_refs'], [test_result_upa]*2)
+        self.assertEquals(len(upas.keys()), 8)
+
+    def assertIsValidCellCode(self, js_obj, data, type, widget, cellId, title):
+        code_lines = js_obj.data.strip().split('\n')
+        self.assertTrue(code_lines[0].strip().startswith('element.html("<div id=\'kb-vis'))
+        self.assertEquals(code_lines[1].strip(), "require(['kbaseNarrativeOutputCell'], function(KBaseNarrativeOutputCell) {")
+        self.assertTrue(code_lines[2].strip().startswith(r"var w = new KBaseNarrativeOutputCell($('#kb-vis"))
+
+
 
 if __name__ == '__main__':
     unittest.main()