+++ test/unit/spec/Util/BootstrapDialogSpec.js
@@ -4,25 +4,19 @@
 /*global beforeEach, afterEach*/
 /*jslint white: true*/
 
-define (
-    [
-        'jquery',
-        'util/bootstrapDialog'
-    ], function(
-        $,
-        Dialog
-    ) {
+define ([
+    'jquery',
+    'util/bootstrapDialog'
+], function(
+    $,
+    Dialog
+) {
+    'use strict';
     var $simpleBody = $('<div>').append('This is a body text'),
         simpleTitle = 'Title',
         simpleButtons = [
-            $('<button>').append('b1')
-                         .click(function() {
-
-                         }),
-            $('<button>').append('b2')
-                         .click(function() {
-
-                         })
+            $('<button>').append('b1').click(function() { }),
+            $('<button>').append('b2').click(function() { })
         ],
         simpleDialog;
 
@@ -61,6 +55,18 @@ define (
             expect(body.html()).toBe($simpleBody.html());
         });
 
+        it('Should give the buttons back', function() {
+            var buttons = simpleDialog.getButtons();
+            expect(buttons.length).toBe(2);
+            expect(buttons[0].innerHTML).toBe('b1');
+            expect(buttons[1].innerHTML).toBe('b2');
+        });
+
+        it('Should get the whole modal object back', function() {
+            var $modal = simpleDialog.getElement();
+            expect($modal[0].tagName).toBe('DIV');
+        });
+
         it('Should write over the title', function() {
             var newTitle = 'A new title!';
             simpleDialog.setTitle(newTitle);
@@ -72,5 +78,54 @@ define (
             simpleDialog.setBody($newBody);
             expect(simpleDialog.getBody().html()).toBe($newBody.html());
         });
+
+        it('Should trigger a command on enter key', function() {
+            var wasTriggered = false;
+            var btnFn = function() {
+                wasTriggered = true;
+            };
+            var dialog = new Dialog({
+                title: 'test',
+                body: $('<div>'),
+                buttons: [$('<button>').append('foo').click(btnFn)],
+                enterToTrigger: true
+            });
+            dialog.show();
+            expect(wasTriggered).toBe(false);
+            var e = $.Event('keypress');
+            e.which = 13;
+            e.keyCode = 13;
+            dialog.getElement().trigger(e);
+            expect(wasTriggered).toBe(true);
+            dialog.hide();
+            dialog.destroy();
+        });
+
+        it('Should create a simple alert', function() {
+            var alert = new Dialog({
+                title: 'Alert',
+                body: $('<div>').append('an alert'),
+                alertOnly: true
+            });
+
+            var btns = alert.getButtons();
+            expect(btns.length).toBe(1);
+            expect(btns[0].innerHTML).toBe('Close');
+        });
+
+        it('Should set the title based on type', function() {
+            var d = new Dialog({
+                title: 'Foo',
+                type: 'warning'
+            });
+            expect(d.$headerTitle.hasClass('text-warning')).toBe(true);
+        });
+
+        it('Should destroy the modal on command', function() {
+            var ret = simpleDialog.destroy();
+            expect(ret).toBe(null);
+            expect(simpleDialog.getElement()).toBe(null);
+        });
+
     });
 });