+++ RELEASE_NOTES.md
@@ -7,7 +7,7 @@ This is built on the Jupyter Notebook v5.4.1 (more notes will follow).
 - Update Jupyter Notebook to 5.4.1 with a few KBase adjustments
     - Prevent Jupyter favicon from overriding ours at various points.
     - Use local version of Font Awesome.
-    - Use local version of Glyphicons font pack (for Datatables-based widgets)
+    - Use local version of Glyphicons font pack (for Datatables-based widgets).
     - Use Bootstrap version 3.3.7
 - Update ipywidgets to 7.2.1
 - SCT-559 - Fix ugly race condition that could prevent app cells from being properly rendered when loading an existing Narrative.
@@ -15,6 +15,10 @@ This is built on the Jupyter Notebook v5.4.1 (more notes will follow).
 - SCT-628 - adds a viewer for the CompoundSet object.
 - Bump Tornado dependency to 5.0.0
 - SCT-637 - adds a warning to the loading section if there's an extreme delay (20 seconds) in between loading individual steps.
+- SCT-690 - truncate long Narrative names, show the whole thing on mouseover.
+- SCT-590 - add cache busting to the public data mapping lookup. No more force-refreshing!
+- SCT-706 - fix problem where space characters were sometimes ignored in the app panel search.
+- Remove old Import tab, remove (New) tag and warning from new Import tab.
 
 ### Version 3.6.3
 - SCT-585 add folder drag and drop upload to the Import area.